refactor(test): Move BuyerOrderControllerTest
